+++ src/main/java/com/uitopic/restock/platform/resources/domain/model/commands/SubtractInventoryCommand.java
@@ -3,7 +3,7 @@ package com.uitopic.restock.platform.resources.domain.model.commands;
 public record SubtractInventoryCommand(
         String branchId,
         String customSupplyId,
-        double quantity,
+        Integer quantity,
         String unit,
         String reason,
         String timestamp
